+++ src/main/java/com/example/demo/controller/GameController/UserDto.java
@@ -1,11 +1,11 @@
 package com.example.demo.controller.GameController;
 
 public class UserDto {
-	public long getUserId() {
+	public int getUserId() {
 		return userId;
 	}
 
-	public void setUserId(long userId) {
+	public void setUserId(int userId) {
 		this.userId = userId;
 	}
 
@@ -17,6 +17,6 @@ public class UserDto {
 		this.userName = userName;
 	}
 
-	private long userId;
+	private int userId;
 	private String userName;
 }
